pyproject.toml
[project]
name = "nginx-access-log-archiver"
version = "1.0.1"
description = "Utility for archiving the department's preserved Nginx access logs in the Parquet column-oriented data file format"
authors = [{ name = "Ashley Felton", email = "ashley.felton@dbca.wa.gov.au" }]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12,<4.0"
dependencies = [
  "python-dotenv==1.1.0",
  "azure-storage-blob==12.25.1",
  "pyarrow==19.0.1",
  "orjson==3.10.16",
  "unicodecsv==0.14.1",
]

[dependency-groups]
dev = ["ipython >= 8.31.0"]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
ignore = [
  "E501", # Line too long
  "E722", # Bare except
]
